Fixed a bug where macros in Plugins>Macros would not be run.
In ImageJA, the Plugins menu is an instance of PopupMenu
instead of Menu.  If MacroInstaller.actionPerformed finds
that the parent menu is a PopupMenu, however, it would
look through each macro to see if there was a "Popup Menu"
macro to run and never run other macros.  This patch will
carry on to run the command if such a macro is not found.


git-svn-id: http://code.imagej.net/svn/imagej/branches/ijx@288 be23795f-a60e-40d4-ba19-055798872e57