Fix Update Menus when the plugins directory is in the classpath
In Fiji, we need to set the classpath to include the classpath, to
be able to access the classes from scripting languages.

But that interferes Update Menus: ImageJ's PluginClassLoader no
longer loads the plugins, and therefore cannot load the new
versions of the classes.

So try loading the classes from the plugins/ folder before
falling back to the system class loader.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>


git-svn-id: http://code.imagej.net/svn/imagej/branches/ijx@621 be23795f-a60e-40d4-ba19-055798872e57