Use the RMI listener by default
When dropping an image onto an icon on Windows or Linux, it is quite
distressing if a new Fiji instance is opened.  This is even more true if
you happen to use all the instances' allocated memory heavily, as that
can pretty much slow down your machine to a halt due to swapping, and
only rebooting helps.

So just reenable the RMI-based listener (which should not have any
security issues due to the use of RMI instead of a plain socket that
is open to all).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>


git-svn-id: http://code.imagej.net/svn/imagej/branches/ijx@768 be23795f-a60e-40d4-ba19-055798872e57