Put back public APIs that have never been marked deprecated
A comment in the source code that a method is obsolete can be missed
easily. Code that is marked @deprecated is less easy to miss, so let's
put back the methods for a certain grace period, to give plugin
developers a chance to modify their source code accordingly.

Suggested by Stephan Saalfeld.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>


git-svn-id: http://code.imagej.net/svn/imagej/branches/ijx@906 be23795f-a60e-40d4-ba19-055798872e57