Add a stub for the macro interpreter
For the moment, only getDimensions() is implemented. But it is a proof of
concept: this function disagrees with the syntax of most Java/C-like
languages as it does not use the values of the passed parameters, but
rather the variable names. It does not even complain when the variables
were not defined previously.

With that task, done we will now implement all of the macro language's
built-in functions in the class imagej.script.MacroFunctions.

Of course, the way we do it will also give plugin programmers all the
macro functions for free, something which was not possible with IJ1.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>