Start to implement macro functions
For the moment, most of the functions are stubs throwing Exceptions, but
the math-related functions are already implemented.

As a proof of concept, bitDepth() is also implemented.

For errors such as when there is no current image, but the currently
executed function requires one, it is nicer to tell the user what is
happening, so we also introduce a new exception, MacroException.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>